Fix text wich gets searched for blocked words
I noticed a mistake on how the text, that is searched through for blocked/censored words with the regex later on, is generated. 

(Line 502 ~ here now 530):
```javascript
var txt = messageData[0];
```
Problem is that messageData is the root of an Object tree.
For example the text input of: `foo **bar** ~~baz :fish: foo *bar **baz***~~` (so "foo **bar** ~~baz :fish: foo *bar **baz***~~" in Discord) is an pretty deeply nested markdown construct with even an emoji inside... currently the plugin does not support markdown at all and would only recognize the word "foo" and then also only the very first one. If it would just be `**foo**` it still would not recognize it...

Fixed by a recursive function (buildText) that flattens out the objects and constructs a string without the markdown and reverses emojis and mentions back to text  (e.g. `:fish:` in the example gets translated back into `:fish:` instead of the pictiure => emojis get searchable)